requirements.txt
flask
pytesseract
matplotlib
opencv-python
numpy
flask-smorest
python-dotenv
flask-sqlalchemy
marshmallow
flask-migrate
flask-jwt-extended
passlib
gunicorn
psycopg2